EE-792: hacky proof of concept reusing existing engine test framework for assemblyscript wasm
