+++ execution-engine/contracts-ts/client/standard-payment/assembly/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "assemblyscript/std/assembly.json",
+  "include": [
+    "./**/*.ts"
+  ]
+}
