--- execution-engine/contracts-ts/standard-payment/assembly/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../../../contract-ffi-ts/node_modules/assemblyscript/std/assembly.json",
-  "include": [
-    "./**/*.ts"
-  ]
-}
