--- execution-engine/contracts-ts/standard-payment/package.json
@@ -1,10 +1,0 @@
-{
-  "scripts": {
-    "asbuild:untouched": "asc --lib ../../ assembly/index.ts -b build/untouched.wasm -t build/untouched.wat --sourceMap --validate --debug --use abort=",
-    "asbuild:optimized": "asc --lib ../../ assembly/index.ts -b build/optimized.wasm -t build/optimized.wat --sourceMap --validate --optimize --use abort=",
-    "asbuild": "npm run asbuild:untouched && npm run asbuild:optimized"
-  },
-  "devDependencies": {
-    "assemblyscript": "^0.8.1"
-  }
-}
